requirements.txt
Django==2.0.7
pytz==2018.5
djangorestframework==3.8.2
gunicorn==19.6.0
Pillow==5.0.0